Rename 'contant' to 'constant' in entity package
Corrects the misspelling of 'contant' to 'constant' in the entity package and updates all relevant imports and documentation references. This improves code clarity and consistency.
