+++ core/src/commonMain/kotlin/work/socialhub/kmisskey/entity/constant/Visibility.kt
@@ -1,4 +1,4 @@
-package work.socialhub.kmisskey.entity.contant
+package work.socialhub.kmisskey.entity.constant
 
 enum class Visibility(
     val code: String
