+++ core/src/commonMain/kotlin/work/socialhub/kmisskey/entity/constant/WebhooksType.kt
@@ -1,4 +1,4 @@
-package work.socialhub.kmisskey.entity.contant
+package work.socialhub.kmisskey.entity.constant
 
 enum class WebhooksType(
     val code: String
